Stop capturing mqtt connection lost errors to sentry
This line was a simple Android log until we introduced SentryLog where it became an exception that got captured. But this case looks very common, no need to spam Sentry with this. This PR stops capturing it into Sentry and only adds it as a breadcrumb like the other mqtt logs.