fix: Sync media files without blocking realm in a transaction
Media synchronization is performed from a realm transaction, which involves a number of operations that take a long time, not to mention that the synchronization itself can take just as long depending on the amount of media stored on the phone.
There are also API calls, which means that the transaction remains blocked for a long time, blocking all other transactions and preventing the application from functioning properly.